// src/api/generated/models/update-user-request-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Users API
 * Documentation of Users API
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface UpdateUserRequestDto
 */
export interface UpdateUserRequestDto {
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequestDto
     */
    'displayedName'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequestDto
     */
    'profilePicture'?: string;
}