// src/api/generated/models/user-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Social API
 * Documentation of Social API
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Details of the user that recieved the friend request
 * @export
 * @interface UserDto
 */
export interface UserDto {
    /**
     * Uniq UUID of the User
     * @type {string}
     * @memberof UserDto
     */
    'id'?: string;
    /**
     * Name that the use chose to display
     * @type {string}
     * @memberof UserDto
     */
    'displayedName'?: string;
    /**
     * Link to the user profile picture
     * @type {string}
     * @memberof UserDto
     */
    'profilePicture'?: string;
    /**
     * Date time of when the user was last online
     * @type {string}
     * @memberof UserDto
     */
    'lastSeen'?: string;
}